+++ pentagon/pentagon.py
@@ -18,8 +18,8 @@ from shutil import copytree, ignore_patterns
 import component.kops as kops
 import component.inventory as inventory
 import component.core as core
+import component.gcp as gcp
 from helpers import render_template, write_yaml_file, create_rsa_key
-from defaults import PentagonDefaults
 from meta import __version__, __author__
 
 
@@ -28,6 +28,85 @@ class PentagonException(Exception):
 
 
 class PentagonProject(object):
+    from defaults import AWSPentagonDefaults as PentagonDefaults
+    keys_to_sanitize = ['aws_access_key', 'aws_secret_key', 'output_file']
+
+    def __init__(self, name, data={}):
+        self._data = data
+        self._name = name
+        logging.debug(self._data)
+
+        self._force = self.get_data('force')
+        self._configure_project = self.get_data('configure')
+
+        # Set this before it gets overridden by the config file
+        self._outfile = self.get_data('output_file')
+
+        # Setting local path info
+        self._repository_name = os.path.expanduser("{}-infrastructure".format(name))
+        self._repository_directory = "{}".format(
+            self._repository_name)
+
+        self._private_path = "inventory/default/config/private"
+
+    def get_data(self, name, default=None):
+        """ Get argument name from click arguments, if it exists, or return default.
+            Builtin .get method is inadequate because click defaults to a value of None
+            which fools the .get() method """
+        if self._data.get(name) is not None:
+            return self._data.get(name)
+        return default
+
+    def __git_init(self):
+        """ Initialize git repository in the project infrastructure path """
+        Repo.init(self._repository_directory)
+
+    def __write_config_file(self):
+        """ Write sanitized yaml file of starting arguments """
+        logging.info("Writing arguments to file for Posterity: {}".format(self._outfile))
+        config = {}
+
+        for key, value in self._data.items():
+            if value and key not in self.keys_to_sanitize:
+                config[key] = value
+
+        logging.debug(config)
+        try:
+            write_yaml_file(self._repository_directory + "/" + self._outfile, config)
+        except Exception as e:
+            logging.debug(traceback.format_exc(e))
+            logging.error("Failed to write arguments to file")
+            logging.error(e)
+
+    def __repository_directory_exists(self):
+        """ Tests if the repository directory already exists """
+        logging.debug("Checking for repository {}".format(self._repository_directory))
+        if os.path.isdir(self._repository_directory):
+            return True
+            logging.debug("Already Exists")
+        logging.debug("Does not exist")
+        return False
+
+    def start(self):
+        if not self.__repository_directory_exists() or self._force:
+            logging.info("Copying project files...")
+            self.__create_repo_core()
+            self.__git_init()
+            self.__write_config_file()
+            with open('{}/.version'.format(self._repository_directory), 'w') as f:
+                f.write(__version__)
+
+            if self._configure_project is not False:
+                self.configure_default_project()
+        else:
+            raise PentagonException('Project path exists. Cowardly refusing to overwrite existing project.')
+
+    def __create_repo_core(self):
+        logging.debug(self._repository_directory)
+        core.Core(self._data).add('{}'.format(self._repository_directory))
+
+
+class AWSPentagonProject(PentagonProject):
     # Placeholders for when there is not sensible default
 
     # AWS and VPC
@@ -62,134 +141,65 @@ class PentagonProject(object):
 
     availability_zone_designations = list(string.ascii_lowercase)
 
-    def get_arg(self, arg_name, default=None):
-        """ Get argument name from click arguments, if it exists, or return default.
-            Builtin .get method is inadequate because click defaults to a value of None
-            which fools the .get() method """
-        if self._args.get(arg_name) is not None:
-            return self._args.get(arg_name)
-        elif os.environ.get('PENTAGON_{}'.format(arg_name), None) is not None:
-            return os.environ.get('PENTAGON_{}'.format(arg_name), None)
-
-        return default
-
-    def __init__(self, name, args={}):
-        self._args = args
-        self._name = name
-        logging.debug(self._args)
-        # Set booleans
-        self._force = args.get('force')
-        self._configure_project = args.get('configure')
-        self._create_keys = args.get('create_keys')
-
-        # Does the git repo already exist
-        self._git_repo = args.get('git_repo')
-
-        # Set this before it gets overridden by the config file
-        self._outfile = self.get_arg('output_file')
-
-        self._config_file = self.get_arg("config_file")
-        if self._config_file is not None:
-            self._args = self.__read_config_file()
-
-        # Setting local path info
-        self._repository_name = os.path.expanduser(self.get_arg("repository_name", "{}-infrastructure".format(name)))
-        self._workspace_directory = os.path.expanduser(self.get_arg('workspace_directory', '.'))
-        self._repository_directory = "{}/{}".format(
-            self._workspace_directory,
-            self._repository_name)
-
-        self._private_path = "inventory/default/config/private"
+    def __init__(self, name, data={}):
+        super(AWSPentagonProject, self).__init__(name, data)
+        self._create_keys = self.get_data('create_keys')
 
         self._ssh_keys = {
-            'admin_vpn_key': self.get_arg('admin_vpn_key', PentagonDefaults.ssh['admin_vpn_key']),
-            'working_kube_key': self.get_arg('working_kube_key', PentagonDefaults.ssh['working_kube_key']),
-            'production_kube_key': self.get_arg('production_kube_key', PentagonDefaults.ssh['production_kube_key']),
-            'working_private_key': self.get_arg('working_private_key', PentagonDefaults.ssh['working_private_key']),
-            'production_private_key': self.get_arg('production_private_key', PentagonDefaults.ssh['production_private_key']),
+            'admin_vpn_key': self.get_data('admin_vpn_key', self.PentagonDefaults.ssh['admin_vpn_key']),
+            'working_kube_key': self.get_data('working_kube_key', self.PentagonDefaults.ssh['working_kube_key']),
+            'production_kube_key': self.get_data('production_kube_key', self.PentagonDefaults.ssh['production_kube_key']),
+            'working_private_key': self.get_data('working_private_key', self.PentagonDefaults.ssh['working_private_key']),
+            'production_private_key': self.get_data('production_private_key', self.PentagonDefaults.ssh['production_private_key']),
         }
 
-        if self._configure_project:
-            # AWS Specific Stuff
-            self._aws_access_key = self.get_arg('aws_access_key', self._aws_access_key_placeholder)
-            self._aws_secret_key = self.get_arg('aws_secret_key', self._aws_secret_key_placeholder)
-            if self.get_arg('aws_default_region'):
-                self._aws_default_region = self.get_arg('aws_default_region')
-                self._aws_availability_zone_count = int(self.get_arg('aws_availability_zone_count', PentagonDefaults.vpc['aws_availability_zone_count']))
-                self._aws_availability_zones = self.get_arg('aws_availability_zones', self.__default_aws_availability_zones())
-            else:
-                self._aws_default_region = self._aws_default_region_placeholder
-                self._aws_availability_zone_count = self._aws_availability_zone_count_placeholder
-                self._aws_availability_zones = self._aws_availability_zones_placeholder
+        # AWS Specific Stuff
+        self._aws_access_key = self.get_data('aws_access_key', self._aws_access_key_placeholder)
+        self._aws_secret_key = self.get_data('aws_secret_key', self._aws_secret_key_placeholder)
+        if self.get_data('aws_default_region'):
+            self._aws_default_region = self.get_data('aws_default_region')
+            self._aws_availability_zone_count = int(self.get_data('aws_availability_zone_count', self.PentagonDefaults.vpc['aws_availability_zone_count']))
+            self._aws_availability_zones = self.get_data('aws_availability_zones', self.__default_aws_availability_zones())
+        else:
+            self._aws_default_region = self._aws_default_region_placeholder
+            self._aws_availability_zone_count = self._aws_availability_zone_count_placeholder
+            self._aws_availability_zones = self._aws_availability_zones_placeholder
 
-            # VPC information
-            self._vpc_name = self.get_arg('vpc_name', PentagonDefaults.vpc['vpc_name'])
-            self._vpc_cidr_base = self.get_arg('vpc_cidr_base', PentagonDefaults.vpc['vpc_cidr_base'])
-            self._vpc_id = self.get_arg('vpc_id', self._vpc_id)
+        # VPC information
+        self._vpc_name = self.get_data('vpc_name', self.PentagonDefaults.vpc['vpc_name'])
+        self._vpc_cidr_base = self.get_data('vpc_cidr_base', self.PentagonDefaults.vpc['vpc_cidr_base'])
+        self._vpc_id = self.get_data('vpc_id', self._vpc_id)
 
-            # DNS
-            self._dns_zone = self.get_arg('dns_zone', '{}.com'.format(self._name))
+        # DNS
+        self._dns_zone = self.get_data('dns_zone', '{}.com'.format(self._name))
 
-            # KOPS:
-            self._infrastructure_bucket = self.get_arg('infrastructure_bucket', self._repository_name)
+        # KOPS:
+        self._infrastructure_bucket = self.get_data('infrastructure_bucket', self._repository_name)
 
-            # Kubernetes version
-            self._kubernetes_version = self.get_arg('kubernetes_version', PentagonDefaults.kubernetes['version'])
+        # Kubernetes version
+        self._kubernetes_version = self.get_data('kubernetes_version', self.PentagonDefaults.kubernetes['version'])
 
-            # Working Kubernetes
-            self._working_kubernetes_cluster_name = self.get_arg('working_kubernetes_cluster_name', 'working-1.{}'.format(self._dns_zone))
-            self._working_kubernetes_dns_zone = self.get_arg('working_kubernetes_dns_zone', '{}'.format(self._dns_zone))
+        # Working Kubernetes
+        self._working_kubernetes_cluster_name = self.get_data('working_kubernetes_cluster_name', 'working-1.{}'.format(self._dns_zone))
+        self._working_kubernetes_dns_zone = self.get_data('working_kubernetes_dns_zone', '{}'.format(self._dns_zone))
 
-            self._working_kubernetes_node_count = self.get_arg('working_kubernetes_node_count', PentagonDefaults.kubernetes['node_count'])
-            self._working_kubernetes_master_aws_zones = self.get_arg('working_kubernetes_master_aws_zones', self._aws_availability_zones)
-            self._working_kubernetes_master_node_type = self.get_arg('working_kubernetes_master_node_type', PentagonDefaults.kubernetes['master_node_type'])
-            self._working_kubernetes_worker_node_type = self.get_arg('working_kubernetes_worker_node_type', PentagonDefaults.kubernetes['worker_node_type'])
-            self._working_kubernetes_v_log_level = self.get_arg('working_kubernetes_v_log_level', PentagonDefaults.kubernetes['v_log_level'])
-            self._working_kubernetes_network_cidr = self.get_arg('working_kubernetes_network_cidr', PentagonDefaults.kubernetes['network_cidr'])
+        self._working_kubernetes_node_count = self.get_data('working_kubernetes_node_count', self.PentagonDefaults.kubernetes['node_count'])
+        self._working_kubernetes_master_aws_zones = self.get_data('working_kubernetes_master_aws_zones', self._aws_availability_zones)
+        self._working_kubernetes_master_node_type = self.get_data('working_kubernetes_master_node_type', self.PentagonDefaults.kubernetes['master_node_type'])
+        self._working_kubernetes_worker_node_type = self.get_data('working_kubernetes_worker_node_type', self.PentagonDefaults.kubernetes['worker_node_type'])
+        self._working_kubernetes_v_log_level = self.get_data('working_kubernetes_v_log_level', self.PentagonDefaults.kubernetes['v_log_level'])
+        self._working_kubernetes_network_cidr = self.get_data('working_kubernetes_network_cidr', self.PentagonDefaults.kubernetes['network_cidr'])
 
-            # Production Kubernetes
-            self._production_kubernetes_cluster_name = self.get_arg('production_kubernetes_cluster_name', 'production-1.{}'.format(self._dns_zone))
-            self._production_kubernetes_dns_zone = self.get_arg('production_kubernetes_dns_zone', '{}'.format(self._dns_zone))
+        # Production Kubernetes
+        self._production_kubernetes_cluster_name = self.get_data('production_kubernetes_cluster_name', 'production-1.{}'.format(self._dns_zone))
+        self._production_kubernetes_dns_zone = self.get_data('production_kubernetes_dns_zone', '{}'.format(self._dns_zone))
 
-            self._production_kubernetes_node_count = self.get_arg('production_kubernetes_node_count', PentagonDefaults.kubernetes['node_count'])
-            self._production_kubernetes_master_aws_zones = self.get_arg('production_kubernetes_master_aws_zones', self._aws_availability_zones)
-            self._production_kubernetes_master_node_type = self.get_arg('production_kubernetes_master_node_type', PentagonDefaults.kubernetes['master_node_type'])
-            self._production_kubernetes_worker_node_type = self.get_arg('production_kubernetes_worker_node_type', PentagonDefaults.kubernetes['worker_node_type'])
-            self._production_kubernetes_v_log_level = self.get_arg('production_kubernetes_v_log_level', PentagonDefaults.kubernetes['v_log_level'])
-            self._production_kubernetes_network_cidr = self.get_arg('production_kubernetes_network_cidr', PentagonDefaults.kubernetes['network_cidr'])
-
-    def __write_config_file(self):
-        logging.info("Writing arguments to file for Posterity: {}".format(self._outfile))
-        config = {}
-
-        if 'output_file' in config:
-            config.pop('output_file')
-
-        for key, value in self._args.items():
-            if value and key != "aws_access_key" and key != "aws_secret_key":
-                config[key] = value
-
-        logging.debug(config)
-        try:
-            write_yaml_file(self._repository_directory + "/" + self._outfile, config)
-        except Exception as e:
-            logging.debug(traceback.format_exc(e))
-            logging.error("Failed to write arguments to file")
-            logging.error(e)
-
-    def __read_config_file(self):
-        logging.info("Reading config file: {}".format(self._config_file))
-        try:
-            with open(self._config_file, "r") as cf:
-                config = yaml.load(cf.read(), Loader=yaml.loader.BaseLoader)
-                logging.debug("Config values: {}".format(config))
-                return config
-        except Exception, e:
-            logging.info("Failed to read arguments from file")
-            logging.error(e)
-            sys.exit(1)
-
-        return {}
+        self._production_kubernetes_node_count = self.get_data('production_kubernetes_node_count', self.PentagonDefaults.kubernetes['node_count'])
+        self._production_kubernetes_master_aws_zones = self.get_data('production_kubernetes_master_aws_zones', self._aws_availability_zones)
+        self._production_kubernetes_master_node_type = self.get_data('production_kubernetes_master_node_type', self.PentagonDefaults.kubernetes['master_node_type'])
+        self._production_kubernetes_worker_node_type = self.get_data('production_kubernetes_worker_node_type', self.PentagonDefaults.kubernetes['worker_node_type'])
+        self._production_kubernetes_v_log_level = self.get_data('production_kubernetes_v_log_level', self.PentagonDefaults.kubernetes['v_log_level'])
+        self._production_kubernetes_network_cidr = self.get_data('production_kubernetes_network_cidr', self.PentagonDefaults.kubernetes['network_cidr'])
 
     def __default_aws_availability_zones(self):
         azs = []
@@ -199,20 +209,8 @@ class PentagonProject(object):
 
         return (", ").join(azs)
 
-    def __repository_directory_exists(self):
-        logging.debug("Verifying repository {}".format(self._repository_directory))
-        if os.path.isdir(self._repository_directory):
-            return True
-        return False
-
-    def __git_init(self):
-        """ Initialize git repository in the project infrastructure path """
-        if self._git_repo:
-            return Git().clone(self._git_repo, self._repository_directory)
-        else:
-            return Repo.init(self._repository_directory)
-
-    def __prepare_context(self):
+    @property
+    def context(self):
         self._context = {
             'aws_secret_key': self._aws_secret_key,
             'aws_access_key': self._aws_access_key,
@@ -226,14 +224,11 @@ class PentagonProject(object):
             'vpc_cidr_base': self._vpc_cidr_base,
             'aws_availability_zones': self._aws_availability_zones,
             'aws_availability_zone_count': self._aws_availability_zone_count,
-            'aws_region': self._aws_default_region,
             'infrastructure_bucket': self._infrastructure_bucket,
             'vpc_name': self._vpc_name,
             'infrastructure_bucket': self._infrastructure_bucket,
             'aws_region': self._aws_default_region,
             'KOPS_STATE_STORE_BUCKET': self._infrastructure_bucket,
-            'org_name': self._name,
-            'vpc_name': self._vpc_name,
             'dns_zone': self._dns_zone,
             'vpn_ami_id': self._vpn_ami_id,
             'production_kube_key': self._ssh_keys['production_kube_key'],
@@ -241,8 +236,10 @@ class PentagonProject(object):
             'production_private_key': self._ssh_keys['production_private_key'],
             'working_private_key': self._ssh_keys['working_private_key'],
             'admin_vpn_key': self._ssh_keys['admin_vpn_key'],
-            'account': 'default',
+            'name': 'default',
         }
+        logging.debug(self._context)
+        return self._context
 
     def __add_kops_working_cluster(self):
         context = {
@@ -288,9 +285,9 @@ class PentagonProject(object):
 
         self._vpn_ami_id = self._vpn_ami_id_placeholder
 
-        if self.get_arg('configure_vpn'):
-            if self.get_arg('vpn_ami_id'):
-                self._vpn_ami_id = self.get_arg('vpn_ami_id')
+        if self.get_data('configure_vpn'):
+            if self.get_data('vpn_ami_id'):
+                self._vpn_ami_id = self.get_data('vpn_ami_id')
             elif \
                     self._aws_access_key != self._aws_access_key_placeholder and \
                     self._aws_secret_key != self._aws_secret_key_placeholder and \
@@ -311,42 +308,54 @@ class PentagonProject(object):
             else:
                 logging.warn("Cannot get ami-id without AWS Key, Secret and Default Region set")
 
-    def start(self):
-
-        if not self.__repository_directory_exists() or self._force:
-            if not self._git_repo:
-                logging.info("Copying project files...")
-                self.__create_repo_core()
-                self.__git_init()
-                self.__configure_default_project()
-                if self._outfile is not None:
-                    self.__write_config_file()
-
-            with open('{}/.version'.format(self._repository_directory), 'w') as f:
-                f.write(__version__)
-        else:
-            raise PentagonException('Project path exists. Cowardly refusing to overwrite existing project.')
-
-    def delete(self):
-        self.__delete()
-
-    def __configure_default_project(self):
+    def configure_default_project(self):
+            self._process_data()
             self.__get_vpn_ami_id()
-            self.__prepare_context()
-
-            inventory.Inventory(self._context).add('{}/inventory/default'.format(self._repository_directory))
-            # self.__prepare_config_private_secrets()
-            # self.__prepare_config_local_vars()
-            # self.__prepare_ssh_config_vars()
-            # self.__prepare_ansible_cfg_vars()
-            # self.__prepare_vpn_cfg_vars()
-            # self.__prepare_account_vars_yml()
-            # self.__prepare_account_vars_sh()
-
+            inventory.Inventory(self.context).add('{}/inventory/default'.format(self._repository_directory))
             self.__add_kops_working_cluster()
             self.__add_kops_production_cluster()
 
-    def __create_repo_core(self):
-        logging.debug(self._repository_directory)
-        core.Core({}).add('{}'.format(self._repository_directory))
-    
+
+class GCPPentagonProject(PentagonProject):
+    from defaults import GCPPentagonDefaults as PentagonDefaults
+    context = {'name': 'default', 'create_keys': False, 'cloud': 'gcp'}
+
+    local_defaults = {
+        'working_cluster_name': 'working',
+        'production_cluster_name': 'production',
+    }
+
+    def __init__(self, name, data={}):
+        super(GCPPentagonProject, self).__init__(name, data)
+        self.default_context = self.PentagonDefaults.kubernetes
+        if type(self.default_context['scopes']) == str:
+            self.default_context['scopes'] = self.default_context['scopes'].split(',')
+        self._data['zones'] = self.get_data('gcp_zones').split(',')
+        self.default_context['project'] = self.get_data('gcp_project')
+
+    def add_working_cluster(self):
+        context = self.default_context
+        context.update({
+            'labels': 'cluster=working',
+            'zone': self.get_data('zones')[0],
+            'node_locations': self.get_data('zones'),
+            'name': self.local_defaults.get('working_cluster_name'),
+            'cluster_ipv4_cidr': self.PentagonDefaults.kubernetes['working_cluster_ipv4_cidr']
+            })
+        gcp.Cluster(context).add('{}/inventory/default/clusters/working'.format(self._repository_directory))
+
+    def add_production_cluster(self):
+        context = self.default_context
+        context.update({
+            'labels': 'cluster=production',
+            'zone': self.get_data('zones')[0],
+            'locations': self.get_data('zones'),
+            'name': self.local_defaults.get('production_cluster_name'),
+            'cluster_ipv4_cidr': self.PentagonDefaults.kubernetes['production_cluster_ipv4_cidr']
+            })
+        gcp.Cluster(context).add('{}/inventory/default/clusters/production'.format(self._repository_directory))
+
+    def configure_default_project(self):
+        inventory.Inventory(self.context).add('{}/inventory/default'.format(self._repository_directory))
+        self.add_working_cluster()
+        self.add_production_cluster()